[day2] Implement day2
I'm glad I didn't try and be clever and work out the values without iterating
over the full range... this code is going to be much easier to adapt for part 2.

I had forgotten to git add the new file when committing part 1, hence the
squashed commit today. It was almost exactly the same with factor being set to 2
and the loop for divisors being unrolled into a single calculation anyway.

[day2] Implement part2

It originally worked for my sample input but not my actual input. I needed to
change the number of digits calculation to use f64 for it to be accurate - I
guess there was a number too big for f32 in there.