[day7] Implement part 1
The splitter index stuff where we translate lazily rather than eagerly looks
kind of nasty but I'm really happy with it, it feels like writing functional
code again.

Part 2 looks like it should be a relatively straightforward update to using a
hashmap and tracking the number of different ways to get to each position rather
than just the set.