[day7] Implement part 2
Super nice, and a great opportunity to use the standard library hashset/hashmap
collections. Always a great feeling when it works first time.

I'm certain there's room to deduplicate the code (just using the part 2 function
with the count for both parts) but I don't mind enough to change it.