[day8] Implement partial part 1
Today is a massive step up in difficulty, although the stage I'm at now is mostly only tricky because I don't know Rust very well, I think.

I was going to do this with Rcs but it's going to be easier for circuits to own all the maps and just index (point lol) into that myself.