[day8] Continue with part 1
I'm so close now... all the actual logic is there and it's just debugging left.

current issue is that I made an assumption every point's next closest point will be the same as the next which I don't think is true in general... though I'm not sure how it's happening.