[day8] Finish part 1, implement part 2
I should have left a commit after part 1 but wanted to minimise my diff. As it
was it was about 5 minutes between finishing part 1 and part 2.

Once I sorted out returning the points in order it just... works! Literally my
first run against the real input after I had something that didn't panic against
the test input just worked first time.

It was an incredible feeling having it work first time as soon as it compiled
honestly.  And it runs in really good time too.

I'll ignore that it still panics for the test input :p